+++ Tests/RelayerTests/AuthTests/ED25519DIDKeyTests.swift
@@ -1,0 +1,14 @@
+import Foundation
+import XCTest
+@testable import WalletConnectRelay
+
+final class ED25519DIDKeyFactoryTests: XCTestCase {
+    let expectedDid = ""
+    var sut: ED25519DIDKeyFactoryImpl!
+
+
+    func test() {
+
+    }
+
+}
