+++ Tests/RelayerTests/AuthTests/SocketAuthenticatorTests.swift
@@ -25,14 +25,3 @@ final class SocketAuthenticatorTests: XCTestCase {
         XCTAssertEqual(token, expectedToken)
     }
 }
-
-//
-//const seed = fromString(
-//  "58e0254c211b858ef7896b00e3f36beeb13d568d47c6031c4218b87718061295",
-//  "base16"
-//);
-//
-//// Generate key pair from seed
-//const keyPair = ed25519.generateKeyPairFromSeed(seed);
-//// secretKey = "58e0254c211b858ef7896b00e3f36beeb13d568d47c6031c4218b87718061295884ab67f787b69e534bfdba8d5beb4e719700e90ac06317ed177d49e5a33be5a"
-//// publicKey = "884ab67f787b69e534bfdba8d5beb4e719700e90ac06317ed177d49e5a33be5a
